clean code: add env and split classes into proper modules. INTEGRATION COMPLETE
