+++ rst-service/src/rst.py
@@ -1,104 +1,21 @@
-#! python
+#!python
 #
 # author: Christian Colonna
 # 
-# classes for rst-service docker api
+# RST classes for rst-service api
 
-import os
-
-from edualigner import RSTEDUAligner
-from PyFred import FREDDialer
-
-import requests
 from nltk.tokenize import word_tokenize
 from rstmarcutree import load_tree as load_rst_file
-from werkzeug.utils import secure_filename
 from rdflib import Graph, URIRef, Literal, XSD
 from rdflib.namespace import Namespace, RDF
+from edualigner import RSTEDUAligner
+from PyFred import FREDDialer
 
 # TODO :
 # this namespace is used by two classes
 # you need to take it from the graph cause
 # g.add() doesn't see it if you didn't declare it
 rst = Namespace('https://rst-ontology-ns/') # rst ontology namespace
-
-
-class FileHandler(object):
-
-    def open_file(self, folder, filename, m='rb'):
-        """ Open file in binary mode. You need this to send it via HTTP POST request.
-        """
-        return open(os.path.join(folder, filename), m)
-
-    def save_secure_file(self, fh, folder):
-        """ Save file received by flask app. Secure it here as arrived from outside.
-        """
-        fh.save(os.path.join(folder, secure_filename(fh.filename)))
-
-
-    def save_file(self, string, folder, filename):
-        """ Save file. 
-        """ 
-        with open(os.path.join(folder, filename), 'w') as fh:
-            fh.write("{}".format(string))
-        fh.close()
-    
-    def set_extension(self, filename, extension):
-        f_root = filename.split(".")[0]
-        return f_root + '.' + extension
-
-    def open_file_to_string(self, folder, filename):
-        fh = open(os.path.join(folder, filename), 'r')
-        raw_file = ''
-        for line in fh:
-            raw_file += line
-        return raw_file    
-
-    def remove_file(self, folder, filename):
-        os.remove(os.path.join(folder, filename))
-    def clean_dir(self, folder, filelist):
-        for filename in filelist:
-            self.remove_file(folder, filename)
-    
-    def parse_request_text(self, request, filekey, folder):
-        """ Check if input is text of a http request is text or file and parse it to a tmp plain_file
-        """
-        if request.files.get(filekey):
-            plain_file = request.files[filekey].filename
-            self.save_secure_file(request.files[filekey], folder)
-        elif request.form.get(filekey):
-            plain_file = "plain.txt"
-            self.save_file(request.form.get(filekey), folder, "plain.txt")
-        return plain_file
-
-
-class DataSender(object):
-    """ Handle data broadcasting over http between docker api. Send files, binaries...
-    """
-
-    def send_file(self, fh, url):
-        return requests.post(url, files={'input': fh})
-
-    def call_rst_service(self, endpoint, in_file, out_file_ext, folder):
-        filehandler = FileHandler()
-        fh = filehandler.open_file(folder, in_file)
-        data = self.send_file(fh, endpoint).text
-        out_file = filehandler.set_extension(in_file, out_file_ext)
-        filehandler.save_file(data, folder, out_file)
-        return out_file
-
-class RDFParams(object):
-    """ RDF parameters for RST model
-    """
-    DOC_NUMBER = 'doc' # this key is the number of the doc to be add to the namespace
-    NAMESPACE = 'ns'
-    DEFAULT_NAMESPACE = 'https://w3id.org/stlab/fred/rst/data/'
-    FORMAT = 'ext'
-    DEFAULT_FORMAT = 'n3'
-    def __init__(self, request_form):
-        self.doc_n = request_form.get(self.DOC_NUMBER) if request_form.get(self.DOC_NUMBER) else 1
-        self.ns = request_form.get(self.NAMESPACE) if request_form.get(self.NAMESPACE) else self.DEFAULT_NAMESPACE
-        self.serialization = request_form.get(self.FORMAT) if request_form.get(self.FORMAT) else self.DEFAULT_FORMAT
 
 class RSTMiner(object):
     """ Take rst tree and produce rdf
@@ -261,27 +178,6 @@ class RSTMiner(object):
             data2send.append({ 'event':row.event, 'agent':row.agent, 'patient':row.patient})
         return { 'important_facts' : data2send }
 
-class GlobalStorage(object):
-
-    def __init__(self):
-        self.raw_text = None
-        self.rdf_params = None
-        self.plain_file = None
-
-    def set_raw_text(self, raw_text):
-        self.raw_text = raw_text
-    def get_raw_text(self):
-        return self.raw_text
-    def set_rdf_params(self, rdf_params):
-        self.rdf_params = rdf_params
-    def get_rdf_params(self):
-        return self.rdf_params    
-    def set_plain_file(self, plain_file):
-        self.plain_file = plain_file
-    def get_plain_file(self):
-        return self.plain_file
-    
-
 class BridgeGraph(Graph):
 
     def merge(self, rst_data, fred_data, in_ext):
@@ -319,7 +215,7 @@ class BridgeGraph(Graph):
         query = "PREFIX rst: <https://rst-ontology-ns/>" \
                 " SELECT  ?textspan WHERE { ?textspan rst:startOffset ?start ; rst:endOffset ?end . FILTER (?start <= ?denoted_start && ?denoted_start < ?end) }"
         qres = self.query(query, initBindings={"?denoted_start":Literal(denoted_tuple[1], datatype=XSD.nonNegativeInteger)})
-#        import pdb; pdb.set_trace()
+        #import pdb; pdb.set_trace()
 
         for row in qres:
             # ============== add membership triple ==================
